feat(middleware): extend deny middleware to deny based on prefix or suffix (#58)
Extends the deny-tag middleware to allow denying based on prefix or suffix.

This PR extends the deny-tag middleware that enables filtering of metric tag values by prefix and/or suffix. The primary use case is to reduce metric cardinality by automatically stripping common prefixes/suffixes from tag values in a centralized way.

The main benefits are:

A central place to filter out tags vs filter it out on the caller side
Reduces the risk of accidentally introducing a high cardinality tag
Opt in/out at a central place
No need to maintain a allow/deny list, tags can be added using a configured format and will automatically be stripped